requirements.txt
pandas
requests
newspaper3k
nltk
openpyxl
lxml-html-clean
tqdm
